requirements.txt
(Libreries)
flask
python-dotenv
pymongo[srv]
gunicorn






























